redis MR4: the outcome — a symbolic error, a transaction, a wait that is not the server's (РR7, РR9, РR10)
PLAN-REDIS.md МR4. МR3 gave an observation its identity and its labels; this
milestone gives it an *outcome*. Three rules, and every one of them exists to
keep a number out of a place where it would be plausible and wrong.

- **the error is symbolic, and two errors are not failures** (РR7). The first
  token of `-WRONGTYPE Operation against a key holding…` is the label; the
  sentence after it is written for a human and quotes the input — the key that
  had the wrong type, the command an ACL refused, the node a `MOVED` points at.
  So src/norm/norm_redis.c grows a second closed table beside the command one
  and folds the token into a pointer *into* it, which is the only form of "no
  byte of a message becomes a label" a test can check rather than trust.
  Case is not folded either: an error symbol is upper-case by convention, and a
  lower-case one is a script's `redis.error_reply('CUSTOMERR …')` — measured in
  `redis/eval-scripts.lkt` — which is exactly what `other` is for.

  `-MOVED` and `-ASK` get `LK_QO_CLIENT_ERR`, the flag a 4xx carries in HTTP,
  and a counter of their own. A resharding cluster answers them continuously
  and every client retries and carries on; in `errors_total` they would paint a
  healthy cluster red for ever. `-MISCONF` and `-LOADING` are the opposite case
  and stay errors — that is a server refusing to work.

- **`+QUEUED` is not a latency** (РR9). Inside a `MULTI` the server writes a
  command down in microseconds and runs it at `EXEC`, so the duration of a
  queued unit measures how fast it can say "noted" and drags every percentile
  down beside real work: `LK_QO_QUEUED`, counted, not timed. The interval that
  does mean something is `MULTI` → the reply to `EXEC`, and it goes to
  `latkit_txn_duration_seconds` unchanged — the one place in this track where a
  cache fits an existing database family exactly.

  The word is read off the *reply* rather than deduced from a `MULTI` we may not
  have seen. That is what makes it right on a connection joined mid-stream, and
  what keeps a queue-time error honest: the offending command is answered
  *instead* of `+QUEUED`, so it keeps its duration and its symbol, and the `EXEC`
  after it is `-EXECABORT`. Two endings need more than the error bit and the
  corpus has both: a broken `WATCH` answers `*-1`/`_` — a **null**, not an error
  — and did not commit, while `*0` is a transaction that ran nothing and did.
  They differ by one byte on the wire and by everything here.

  Measured while writing the tests and now in the notes: after a refused nested
  `MULTI`, Redis 7.4 keeps the transaction and answers the next `EXEC` with `*0`,
  Valkey 8 treats the refusal as a queue-time error and answers `-EXECABORT`.
  Both are read correctly precisely because the verdict comes from the reply and
  not from a model of what a server ought to do — the same reason РR5/РR6's
  labels move on the reply.

- **a blocking command measures the client's own patience** (РR10). `BLPOP key
  30` is a thirty-second observation about a third party's traffic; one of them
  in the general histogram decides the p99 of the whole instance. `LK_QO_BLOCKING`
  keeps it out (МR5 gives it `latkit_redis_blocking_seconds`).

  `XREAD`/`XREADGROUP` are the two the server's own flag gets wrong — they block
  only with a `BLOCK` keyword — and refining that is **the only place in the tree
  an argument past the identity is read at all**, so the bound is stated three
  ways: only for those two commands (their own table bit, so nothing else pays
  for the deeper argv read), only for the keyword's presence and never its value,
  and only before `STREAMS`, after which every element is a key and a stream may
  perfectly well be called `BLOCK`.

The metrics facade acts on the two new flags today (a flag nothing reads is a
flag that is not tested); the branch is marked for МR5 to move into the redis
profile. The `--queries` view grows `err=`, the stats line `redirects=`.

Tests: test_redis_error.c, test_redis_txn.c and test_redis_block.c drive the
real chain — bytes → framer → lk_msg → handler → observation — and the trace
acceptance is now the МR4 half too: on `cluster/moved` the `MOVED`s are in the
redirects and in no error counter while the one `CROSSSLOT` is the other way
round; on `redis/multi` six intervals, nine `+QUEUED` units, and seventeen of
twenty-six observations in the duration histogram; on `redis/blocking` five of
eight kept out of it. redis_privacy.sh gains the two error-heavy traces, since
an error message is a new surface that quotes keys and node addresses.

Also fixed on the way past: the fuzz profile has not linked since МR1 — its
source list never learned about src/proto/redis, so all five targets failed on
`lk_proto_redis_ops`. `fuzz_redis` itself is still МR8's.

+++ tests/unit/test_redis_cmd.c
@@ -461,7 +461,7 @@ static int test_closed_set(void)
 static void dump(void)
 {
     for (uint32_t id = 0; id < lk_redis_cmd_count(); id++) {
-        uint16_t f = lk_redis_cmd_flags(id);
+        uint32_t f = lk_redis_cmd_flags(id);
         uint32_t len;
         const char *name = lk_redis_cmd_name(id, &len);
         char bits[4];
